Bug with Clustering CustomPointMarkers fixed
