+++ CoreMapView/nbproject/project.xml
@@ -194,6 +194,8 @@
                 <package>au.gov.asd.tac.constellation.views.mapview.overlays</package>
                 <package>au.gov.asd.tac.constellation.views.mapview.providers</package>
                 <package>au.gov.asd.tac.constellation.views.mapview.utilities</package>
+                <package>au.gov.asd.tac.constellation.views.mapview2</package>
+                <package>au.gov.asd.tac.constellation.views.mapview2.plugins</package>
             </public-packages>
         </data>
     </configuration>
